Main Player Code
This is our Final version of the "main code". The previous versions of this come from "THoldEm(may5th).py", "THoldEm(may3rd).py, and "THoldEm.py". In this version we have implemented the UDP socket programming, some added functionality to the GUI (using pygame), and fixed some logic errors with the Logic of this poker game (i.e. folding in the middle of a betting round). REMEMBER using UDP requires you to change the "host" ip address to appropriate ipv4 (internet protocal version 4) accessible IP.