Final Second Player Code
This is a "dummy player" we created to just display player two's information to the screen using pygame. So, it doesn't make any calculations but instead receives the calculations from the "MainPlayer_FINALFINAL.py" code using UDP socket programming and then displays it using pygame. Note that we only call the UDP_client() function in this code.